Refactor component selection page
Component selection page has a lot of widgets compared to other pages so
moved the implementation to own file. Also fixed the following issues
in component selection page:
1. Fixed crash when there were no categories and QBsp button was pressed
2. QBsp button opened multiple dialogs if next/back buttons were pressed
3. Made progressbar use the entire vertical space
