+++ src/libs/installer/installer.pro
@@ -132,7 +132,8 @@ HEADERS += packagemanagercore.h \
     lib7z_create.h \
     lib7z_extract.h \
     lib7z_list.h \
-    repositorycategory.h
+    repositorycategory.h \
+    componentselectionpage_p.h
 
 SOURCES += packagemanagercore.cpp \
     packagemanagercore_p.cpp \
@@ -208,7 +209,8 @@ SOURCES += packagemanagercore.cpp \
     keepaliveobject.cpp \
     systeminfo.cpp \
     packagesource.cpp \
-    repositorycategory.cpp
+    repositorycategory.cpp \
+    componentselectionpage_p.cpp
 
 FORMS += proxycredentialsdialog.ui \
     serverauthenticationdialog.ui
